Create face_mask_detection_system.ipynb
Detection of a face masks in a webcam stream the flow to determine whether or not the person in the webcam is wearing a face mask. There are two steps to the procedure:
1) To recognize the people in the webcam. Sort the faces into groups based on their masks.
2) A pre-trained model from the OpenCV framework was used to recognize the faces. Web photographs were used to train the model.
We will use OpenCV to detect faces in real time from a live broadcast from our webcam.
The model has been taught to recognize faces and determine whether or not the person is wearing a mask.
Face photos are required for our mask detection technology to function. We'll use the methods discussed above to detect frames with faces for this. To begin, we need to import the appropriate libraries.